Updated Ping Flag
The wait time flag when running the if statement `if ! ping -c 1 -w 5 $rvHost > /dev/null 2>&1 ; then` should be capitalized. Once i made the change `if ! ping -c 1 -W 5 $rvHost > /dev/null 2>&1 ; then` my host was able to be reached. I also tested it agaisnt a false host and it gave the error message like it is supposed to.